+++ src/txdb.h
@@ -94,7 +94,7 @@ public:
 
     bool GetKey(COutPoint &key) const override;
     bool GetValue(Coin &coin) const override;
-    unsigned int GetValueSize() const override;
+    size_t GetValueSize() const override;
 
     bool Valid() const override;
     void Next() override;
